refactor(communities):  Improving Community component code organization
